Issue #28  / [#82749292] fix Incident report from Unit Correspondent maintenance doc.
