+++ src/main/resources/org/kuali/kra/datadictionary/docs/IacucUnitCorrespondentMaintenanceDocument.xml
@@ -31,7 +31,8 @@
             <list>
                 <bean parent="MaintainableFieldDefinition" p:name="personId" p:required="true" />
                 <!-- ### Vivantech Fix : #25 / [#82749864] fix the issue for no return value from KC Person lookup -->
- 		        <bean parent="MaintainableFieldDefinition" p:name="person.userName" p:required="false" p:overrideLookupClass="org.kuali.kra.bo.KcPerson" p:overrideFieldConversions="personId:personId"/>
+                <!-- ### Vivantech Fix : #28 / [#82749292] KC Person should be required field -->
+ 		        <bean parent="MaintainableFieldDefinition" p:name="person.userName" p:required="true" p:overrideLookupClass="org.kuali.kra.bo.KcPerson" p:overrideFieldConversions="personId:personId"/>
                 <bean parent="MaintainableFieldDefinition" p:name="unitNumber" p:required="true" />
                 <bean parent="MaintainableFieldDefinition" p:name="correspondentTypeCode" p:required="true" />
                 <bean parent="MaintainableFieldDefinition" p:name="nonEmployeeFlag" p:required="true" />
